requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
faiss-cpu==1.7.4
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.5.0
huggingface-hub==0.14.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.3
openai==0.27.7
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.29.2
typing_extensions==4.6.2
urllib3==2.0.2
yarl==1.9.2
